+++ src/hittable.rs
@@ -3,17 +3,17 @@ use nalgebra::Vector3;
 
 #[derive(Clone)]
 pub struct HitRecord {
-    pub point: Vector3<f32>,
-    pub normal: Vector3<f32>,
-    pub t: f32,
+    pub point: Vector3<f64>,
+    pub normal: Vector3<f64>,
+    pub t: f64,
     pub front: bool,
 }
 
 impl HitRecord {
     pub fn new() -> Self {
         Self {
-            point: Vector3::<f32>::zeros(),
-            normal: Vector3::<f32>::zeros(),
+            point: Vector3::<f64>::zeros(),
+            normal: Vector3::<f64>::zeros(),
             t: 0.0,
             front: false,
         }
@@ -21,8 +21,8 @@ impl HitRecord {
 }
 
 pub trait Hittable {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord>;
-    fn normal(&self, p: &Vector3<f32>) -> Vector3<f32>;
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
+    fn normal(&self, p: &Vector3<f64>) -> Vector3<f64>;
     fn clone_dyn(&self) -> Box<dyn Hittable>;
 }
 
@@ -50,7 +50,7 @@ impl HitList {
         self.objects.push(object);
     }
 
-    pub fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
+    pub fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
         let mut first_hit: Option<HitRecord> = None;
 
         self.objects
